several refactors
